refactor: move path variables in constants
