--- config/paths.js
@@ -1,9 +1,0 @@
-const IMAGES_PATH = `./src/assets/images`
-const POSTS_PATH = `./posts`
-const POST_TEMPLATE_PATH = `./src/templates/post/index.js`
-
-module.exports = {
-  IMAGES_PATH,
-  POSTS_PATH,
-  POST_TEMPLATE_PATH,
-}
